+++ Assets/Xbox Live/Editor/XboxLivePostProcessingUnityProject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4fe701e19650d9541b767bce400e693e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
